% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{boot_ci}
\alias{boot_ci}
\title{Confidence Intervals for bootstrapped vectors}
\usage{
boot_ci(x)
}
\arguments{
\item{x}{A vector.}
}
\value{
The lower and upper confidence intervals of \code{x}.
}
\description{
Compute confidence intervals for a vector of bootstrapped values.
}
\examples{
data(efc)
bs <- bootstrap(efc, 100)

# now run models for each bootstrapped sample
bs$models <- lapply(bs$strap, function(x) lm(neg_c_7 ~ e42dep + c161sex, data = x))

# extract coefficient "dependency" from each model
bs$dependency <- unlist(lapply(bs$models, function(x) coef(x)[2]))

# get bootstrapped confidence intervals
boot_ci(bs$dependency)

# compare with model fit
fit <- lm(neg_c_7 ~ e42dep + c161sex, data = efc)
confint(fit)[2, ]

# compare coefficients
mean(bs$dependency)
coef(fit)[2]

}

