% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chi_squared_test.R
\name{chi_squared_test}
\alias{chi_squared_test}
\title{Chi-Squared Test}
\usage{
chi_squared_test(
  data,
  select = NULL,
  by = NULL,
  probabilities = NULL,
  weights = NULL,
  ...
)
}
\arguments{
\item{data}{A data frame.}

\item{select}{The dependent variable (numeric) to be used for the test.}

\item{by}{The grouping variable (factor) to be used for the test. If \code{by} is
not a factor, it will be coerced to a factor.}

\item{weights}{An optional weighting variable (numeric) to be used for the test.}

\item{...}{Additional arguments passed down to \code{\link[=chisq.test]{chisq.test()}}.}
}
\value{
A data frame.
}
\description{
This function performs a Mann-Whitney-Test (or Wilcoxon rank
sum test for \emph{unpaired} samples, see \code{\link[=wilcox.test]{wilcox.test()}} and \code{\link[coin:LocationTests]{coin::wilcox_test()}}).

The function reports p and Z-values as well as effect size r and group-rank-means.
}
\details{
This function calls \code{\link[coin:LocationTests]{coin::wilcox_test()}} to extract effect sizes.
Interpretation of the effect size \strong{r}, as a rule-of-thumb:
\itemize{
\item small effect >= 0.1
\item medium effect >= 0.3
\item large effect >= 0.5
}

\strong{r} is calcuated as:

\if{html}{\out{<div class="sourceCode">}}\preformatted{r = |Z| / sqrt(n1 + n2)
}\if{html}{\out{</div>}}
}
\examples{
data(efc)
# Mann-Whitney-U-Tests for elder's age by elder's sex.
chi_squared_test(efc, "e17age", "e16sex")
}
