% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cronb.R
\name{cronb}
\alias{cronb}
\title{Cronbach's Alpha for a matrix or data frame}
\usage{
cronb(data)
}
\arguments{
\item{data}{\code{data.frame} or matrix with more than 2 columns.}
}
\value{
The Cronbach's alpha value for \code{data}.
}
\description{
This function calculates the Cronbach's alpha value
               of a data frame or matrix.
}
\note{
See 'Examples' from \code{\link[sjPlot]{sjp.pca}} and \code{\link[sjPlot]{sjt.pca}}.
}
\seealso{
\code{\link{reliab_test}}
}

