% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merMod_p.R
\name{get_model_pval}
\alias{get_model_pval}
\alias{merMod_p}
\title{Get p-values from regression model objects}
\usage{
get_model_pval(fit, p.kr = FALSE)

merMod_p(fit, p.kr = TRUE)
}
\arguments{
\item{fit}{A fitted model object of \code{lm}, \code{glm}, \code{merMod},
\code{merModLmerTest}, \code{pggls} or \code{gls}. Other classes may
work as well.}

\item{p.kr}{Logical, if \code{TRUE}, the computation of p-values is based on
conditional F-tests with Kenward-Roger approximation for the df (see
'Details').}
}
\value{
A \code{\link{tibble}} with the model coefficients' names (\code{term}),
        p-values (\code{p.value}) and standard errors (\code{std.error}).
}
\description{
This function returns the p-values for fitted model objects.
}
\details{
For linear mixed models (\code{lmerMod}-objects), the computation of
        p-values (if \code{p.kr = TRUE}) is based on conditional F-tests
        with Kenward-Roger approximation for the df, using the
        \CRANpkg{pbkrtest}-package. If \pkg{pbkrtest} is not available or
        \code{p.kr = FALSE}, or if \code{x} is a \code{glmerMod}-object,
        computation of p-values is based on normal-distribution assumption,
        treating the t-statistics as Wald z-statistics.
        \cr \cr
        If p-values already have been computed (e.g. for \code{merModLmerTest})-objects
        from the \CRANpkg{lmerTest}-package), these will be returned.
}
\note{
\code{merMod_p()} will become deprecated soon, please use
      \code{get_model_pval()} instead.
}
\examples{
data(efc)
# linear model fit
fit <- lm(neg_c_7 ~ e42dep + c172code, data = efc)
get_model_pval(fit)

# Generalized Least Squares fit
library(nlme)
fit <- gls(follicles ~ sin(2*pi*Time) + cos(2*pi*Time), Ovary,
           correlation = corAR1(form = ~ 1 | Mare))
get_model_pval(fit)

# lme4-fit
library(lme4)
fit <- lmer(Reaction ~ Days + (Days | Subject), data = sleepstudy)
get_model_pval(fit, p.kr = TRUE)

}

