% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_prime.R
\name{is_prime}
\alias{is_prime}
\title{Find prime numbers}
\usage{
is_prime(x)
}
\arguments{
\item{x}{An integer, or a vector of integers.}
}
\value{
\code{TRUE} for each prime number in \code{x}, \code{FALSE} otherwise.
}
\description{
This functions checks whether a number is, or numbers in a
vector are prime numbers.
}
\examples{
is_prime(89)
is_prime(15)
is_prime(c(1, 2, 3, 4, 5, 6, 7, 8, 9, 10))

}
