% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kruskal_wallis_test.R
\name{kruskal_wallis_test}
\alias{kruskal_wallis_test}
\title{Kruskal-Wallis test}
\usage{
kruskal_wallis_test(data, select = NULL, by = NULL, weights = NULL)
}
\arguments{
\item{data}{A data frame.}

\item{select}{Name(s) of the continuous variable(s) (as character vector)
to be used as samples for the test. \code{select} can be one of the following:
\itemize{
\item \code{select} specifies \strong{one} variable and \code{by} is provided to indicate the
groups of comparison. In this case, the samples in \code{select} are split by
the groups in \code{by}.
\item If \code{select} specifies \strong{one} variable and \code{by = NULL}, a one-sample test
is carried out (only applicable for \code{t_test()} and \code{wilcoxon_test()}).
\item If \code{select} specifies \strong{two} variables and \code{by} is \code{NULL}, the two
samples are compared. In combination with \code{paired}, a paired test is
carried out, i.e. samples are considered as \emph{dependent}.
}}

\item{by}{Name of the variable indicating the groups. Required if \code{select}
specifies only one variable that contains all samples to be compared in the
test. If \code{by} is not a factor, it will be coerced to a factor. For
\code{chi_squared_test()}, if \code{probabilities} is provided, \code{by} must be \code{NULL}.}

\item{weights}{Name of an (optional) weighting variable to be used for the test.}
}
\value{
A data frame with test results.
}
\description{
This function performs a Kruskal-Wallis rank sum test, to test
the null hypothesis that the population median of all of the groups are
equal. The alternative is that they differ in at least one. Unlike the
underlying base R function \code{kruskal.test()}, this function allows for
weighted tests.
}
\details{
The function simply is a wrapper around \code{\link[=kruskal.test]{kruskal.test()}}. The
weighted version of the Kruskal-Wallis test is based on the \strong{survey} package,
using \code{\link[survey:svyranktest]{survey::svyranktest()}}.
}
\examples{
data(efc)
# Kruskal-Wallis test for elder's age by education
kruskal_wallis_test(efc, "e17age", by = "c172code")

# when data is in wide-format, specify all relevant continuous
# variables in `select` and omit `by`
set.seed(123)
wide_data <- data.frame(
  scale1 = runif(20),
  scale2 = runif(20),
  scale3 = runif(20)
)
kruskal_wallis_test(wide_data, select = c("scale1", "scale2", "scale3"))

# same as if we had data in long format, with grouping variable
long_data <- data.frame(
  scales = c(wide_data$scale1, wide_data$scale2, wide_data$scale3),
  groups = rep(c("A", "B", "C"), each = 20)
)
kruskal_wallis_test(long_data, select = "scales", by = "groups")
# base R equivalent
kruskal.test(scales ~ groups, data = long_data)
}
\seealso{
\itemize{
\item \code{\link[=mann_whitney_test]{mann_whitney_test()}} for unpaired (independent) samples.
\item \code{\link[=t_test]{t_test()}} for parametric t-tests.
\item \code{\link[=kruskal_wallis_test]{kruskal_wallis_test()}} for non-parametric ANOVA (i.e. more than two samples).
\item \code{\link[=wilcoxon_test]{wilcoxon_test()}} for Wilcoxon rank sum tests for paired (dependent) samples.
\item \code{\link[=chi_squared_test]{chi_squared_test()}} for chi-squared tests (two categorical variables).
}
}
