% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sjStatistics.R
\name{levene_test}
\alias{levene_test}
\title{Levene-Test for One-Way-Anova}
\usage{
levene_test(depVar, grpVar)
}
\arguments{
\item{depVar}{Dependent variable.}

\item{grpVar}{Grouping (independent) variable, as unordered factor.}
}
\description{
Plot results of Levene's Test for Equality of Variances for One-Way-Anova.
}
\examples{
data(efc)
levene_test(efc$c12hour, efc$e42dep)

}
