% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_link_inverse.R
\name{link_inverse}
\alias{link_inverse}
\title{Get inverse link function from model objects}
\usage{
link_inverse(model)
}
\arguments{
\item{model}{A model object.}
}
\value{
If known, the inverse link function from \code{model}; else \code{NULL}
        for those models where the inverse link function can't be identified.
}
\description{
Get inverse link function from model objects.
}
\examples{
data(efc)
m <- lm(neg_c_7 ~ e42dep + c172code, efc)
link_inverse(m)(2.3)

# example from ?stats::glm
counts <- c(18, 17, 15, 20, 10, 20, 25, 13, 12)
outcome <- gl(3, 1, 9)
treatment <- gl(3, 3)
m <- glm(counts ~ outcome + treatment, family = poisson())

link_inverse(m)(.3)
# same as
exp(.3)


}
