% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mann_whitney_test.R
\name{mann_whitney_test}
\alias{mann_whitney_test}
\title{Mann-Whitney test}
\usage{
mann_whitney_test(
  data,
  select = NULL,
  by = NULL,
  weights = NULL,
  distribution = "asymptotic"
)
}
\arguments{
\item{data}{A data frame.}

\item{select}{One or more name of the continuous variable (as character
vector) to be used as samples for the test. If \code{select} only specified one
variable, a one-sample test is carried out (only applicable for \code{t_test()}).
Else, \code{by} must be provided to indicate the groups of comparison.}

\item{by}{Name of the variable indicating the groups. Required if \code{select}
specifies only one variable that contains all samples to be compared in the
test. If \code{by} is not a factor, it will be coerced to a factor. For
\code{chi_squared_test()}, if \code{probabilities} is provided, \code{by} must be \code{NULL}.}

\item{weights}{Name of an (optional) weighting variable to be used for the test.}

\item{distribution}{Indicates how the null distribution of the test statistic
should be computed. May be one of \code{"exact"}, \code{"approximate"} or \code{"asymptotic"}
(default). See \code{\link[coin:LocationTests]{coin::wilcox_test()}} for details.}
}
\value{
A data frame with test results. The function returns p and Z-values
as well as effect size r and group-rank-means.
}
\description{
This function performs a Mann-Whitney test (or Wilcoxon rank
sum test for \emph{unpaired} samples.

A Mann-Whitney test is a non-parametric test for the null hypothesis that two
independent samples have identical continuous distributions. It can be used
when the two continuous variables are not normally distributed.
}
\details{
This function is based on \code{\link[=wilcox.test]{wilcox.test()}} and \code{\link[coin:LocationTests]{coin::wilcox_test()}}
(the latter to extract effect sizes). The weighted version of the test is
based on \code{\link[survey:svyranktest]{survey::svyranktest()}}.

Interpretation of the effect size \strong{r}, as a rule-of-thumb:
\itemize{
\item small effect >= 0.1
\item medium effect >= 0.3
\item large effect >= 0.5
}

\strong{r} is calcuated as:

\if{html}{\out{<div class="sourceCode">}}\preformatted{r = |Z| / sqrt(n1 + n2)
}\if{html}{\out{</div>}}
}
\examples{
\dontshow{if (requireNamespace("coin", quietly = TRUE) && requireNamespace("survey", quietly = TRUE)) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
data(efc)
# Mann-Whitney-U tests for elder's age by elder's sex.
mann_whitney_test(efc, "e17age", by = "e16sex")

# when data is in wide-format, specify all relevant continuous
# variables in `select` and omit `by`
set.seed(123)
wide_data <- data.frame(scale1 = runif(20), scale2 = runif(20))
mann_whitney_test(wide_data, select = c("scale1", "scale2"))

# same as if we had data in long format, with grouping variable
long_data <- data.frame(
  scales = c(wide_data$scale1, wide_data$scale2),
  groups = rep(c("A", "B"), each = 20)
)
mann_whitney_test(long_data, select = "scales", by = "groups")
\dontshow{\}) # examplesIf}
}
