% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcse.R, R/n_eff.R
\name{mcse}
\alias{mcse}
\alias{mcse.stanmvreg}
\alias{mcse.stanreg}
\alias{n_eff}
\alias{n_eff.stanreg}
\alias{n_eff.brmsfit}
\title{Compute statistics for MCMC samples and Stan models}
\usage{
mcse(x, ...)

\method{mcse}{stanmvreg}(x, type = c("fixed", "random", "all"), ...)

\method{mcse}{stanreg}(x, type = c("fixed", "random", "all"), ...)

n_eff(x, ...)

\method{n_eff}{stanreg}(x, type = c("fixed", "random", "all"), ...)

\method{n_eff}{brmsfit}(x, type = c("fixed", "random", "all"), ...)
}
\arguments{
\item{x}{A \code{stanreg}, \code{stanfit}, or \code{brmsfit} object.}

\item{...}{Not used}

\item{type}{For mixed effects models, specify the type of effects that should
be returned. \code{type = "fixed"} returns fixed effects only,
\code{type = "random"} the random effects and \code{type = "all"} returns
both fixed and random effects.}
}
\value{
\code{mcse()} and \code{n_eff()} return a tibble with two columns: one
  with the term names and one with the related statistic resp. effective
  sample size.
}
\description{
\code{n_eff()} calculates the the number of effective samples
  (effective sample size). \code{mcse()} returns the Monte Carlo standard
  error.
}
\details{
\describe{
    \item{\strong{MCSE}}{
      The Monte Carlo Standard Error is another useful measure of accuracy of
      the chains. It is defined as standard deviation of the chains divided by
      their effective sample size (the formula for \code{mcse()} is from
      Kruschke 2015, p. 187). The MCSE \dQuote{provides a quantitative suggestion
      of how big the estimation noise is}.
    }
    \item{\strong{Number of Effective Samples}}{
      The effective sample size divides the actual sample size by the amount
      of autocorrelation. The effective sample size is a measure of \dQuote{how
      much independent information there is in autocorrelated chains}, or:
      \dQuote{What would be the sample size of a completely non-autocorrelated chain
      that yielded the same information?} (\emph{Kruschke 2015, p182-3}).
      The ratio of effective number of samples and total number of samples
      (provided in \code{tidy_stan()}) ranges from 0 to 1, and should be close
      to 1. The closer this ratio comes to zero means that the chains may be
      inefficient, but possibly still okay.
    }
  }
}
\examples{
\dontrun{
if (require("rstanarm")) {
  m <- stan_glm(mpg ~ wt + am, data = mtcars, chains = 1)
  mcse(m)
  n_eff(m)
}}
}
\references{
Kruschke JK. Doing Bayesian Data Analysis: A Tutorial with R, JAGS, and Stan. 2nd edition. Academic Press, 2015
}
