% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mic.R
\name{mic}
\alias{mic}
\title{Mean Inter-Item-Correlation}
\usage{
mic(data, cor.method = c("pearson", "spearman", "kendall"))
}
\arguments{
\item{data}{A \code{matrix} as returned by the \code{\link{cor}}-function, or
a \code{data.frame}, where correlations will be calculated.}

\item{cor.method}{Indicates the correlation computation method. May be one of
\code{"spearman"} (default), \code{"pearson"} or \code{"kendall"}.
You may use initial letter only.}
}
\value{
The value of the computed mean inter-item-correlation.
}
\description{
This function calculates a mean inter-item-correlation, i.e.
               a correlation matrix of \code{data} will be computed (unless
               \code{data} is already a matrix as returned by the
               \code{\link{cor}}-function) and the mean
               of the sum of all item's correlation values is returned.
               Requires either a data frame or a computed \code{\link{cor}}-object.
}
\note{
\dQuote{Ideally, the average inter-item correlation for a set of
         items should be between .20 and .40, suggesting that while the
         items are reasonably homogenous, they do contain sufficiently
         unique variance so as to not be isomorphic with each other.
         When values are lower than .20, then the items may not be
         representative of the same content domain. If values are higher than
         .40, the items may be only capturing a small bandwidth of the construct.}
         \cite{(Piedmont 2014)}
}
\examples{
# Data from the EUROFAMCARE sample dataset
data(efc)
# recveive first item of COPE-index scale
start <- which(colnames(efc) == "c82cop1")
# recveive last item of COPE-index scale
end <- which(colnames(efc) == "c90cop9")
# create data frame with COPE-index scale
mydat <- data.frame(efc[, c(start:end)])

mic(mydat)

}
\references{
Piedmont RL (2014) Inter-item Correlations. In: Michalos AC (eds)
            Encyclopedia of Quality of Life and Well-Being Research.
            Dordrecht: Springer, 3303-3304
            \doi{10.1007/978-94-007-0753-5_1493}
}

