% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sum.R
\name{mn}
\alias{mn}
\alias{sm}
\alias{md}
\title{Sum, mean and median for vectors}
\usage{
mn(x, na.rm = TRUE)

sm(x, na.rm = TRUE)

md(x, na.rm = TRUE)
}
\arguments{
\item{x}{A vector.}

\item{na.rm}{Logical, whether to remove NA values from \code{x} before computing
mean, median or sum.}
}
\value{
The mean, median or sum of \code{x}.
}
\description{
\code{mn()}, \code{md()} and \code{sm()} calculate the mean,
       median or sum of values in a vector, but have set argument \code{na.rm}
       to \code{TRUE} by default.
}
\examples{
data(efc)
md(efc$neg_c_7)
mn(efc$neg_c_7)
mean(efc$neg_c_7)
sm(efc$neg_c_7 > 15)

}
