% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phi.R
\name{phi}
\alias{phi}
\title{Phi value for contingency tables}
\usage{
phi(tab)
}
\arguments{
\item{tab}{A \code{\link{table}} or \code{\link{ftable}}. Tables of class
\code{\link{xtabs}} and other will be coerced to \code{\link{ftable}}
objects.}
}
\value{
The table's Phi value.
}
\description{
Compute Phi value for a contingency table.
}
\examples{
tab <- table(sample(1:2, 30, TRUE), sample(1:2, 30, TRUE))
phi(tab)

}
\seealso{
\code{\link{cramer}}
}

