% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Deprecated.R
\name{r2}
\alias{r2}
\alias{cohens_f}
\alias{eta_sq}
\alias{epsilon_sq}
\alias{omega_sq}
\alias{scale_weights}
\alias{robust}
\alias{icc}
\alias{p_value}
\alias{se}
\alias{means_by_group}
\alias{mean_n}
\title{Deprecated functions}
\usage{
r2(x)

cohens_f(x, ...)

eta_sq(x, ...)

epsilon_sq(x, ...)

omega_sq(x, ...)

scale_weights(x, ...)

robust(x, ...)

icc(x)

p_value(x, ...)

se(x, ...)

means_by_group(x, ...)

mean_n(x, ...)
}
\arguments{
\item{x}{An object.}

\item{...}{Currently not used.}
}
\value{
Nothing.
}
\description{
A list of deprecated functions.
}
