% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/re-exports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{\%>\%}
\alias{typical_value}
\alias{mse}
\alias{rmse}
\alias{ci}
\alias{equivalence_test}
\alias{mediation}
\alias{link_inverse}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{bayestestR}{\code{\link[bayestestR]{ci}}, \code{\link[bayestestR]{equivalence_test}}, \code{\link[bayestestR]{mediation}}}

  \item{insight}{\code{\link[insight]{link_inverse}}}

  \item{magrittr}{\code{\link[magrittr:pipe]{\%>\%}}}

  \item{performance}{\code{\link[performance:performance_mse]{mse}}, \code{\link[performance:performance_rmse]{rmse}}}

  \item{sjmisc}{\code{\link[sjmisc]{typical_value}}}
}}

