\encoding{UTF-8}
\name{sjstats-package}
\alias{sjstats-package}
\alias{sjstats}
\docType{package}
\title{Data Transformation and Labelled Data Utility Functions}
\description{
This package is a collection of functions that help computing various statistics
(mostly) for simple regression or mixed models. Some functions are also for
simple descriptive statistics.
}
\details{
\tabular{rl}{
Package: \tab sjstats\cr
Type: \tab Package\cr
Version: \tab 0.1.0\cr
Date: \tab 2016-05-20\cr
License: \tab GPL-3\cr
}
}
\author{
Daniel Lüdecke \email{d.luedecke@uke.de}
}
\references{
  \itemize{
    \item Documentation: \url{http://www.strengejacke.de/sjPlot}
    \item Github: \url{https://github.com/sjPlot/sjstats}
    \item Bug reports: \url{https://github.com/sjPlot/sjstats/issues}
    }
}
