% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_test.R
\name{t_test}
\alias{t_test}
\title{Student's t test}
\usage{
t_test(
  data,
  select = NULL,
  by = NULL,
  weights = NULL,
  paired = FALSE,
  mu = 0,
  alternative = "two.sided"
)
}
\arguments{
\item{data}{A data frame.}

\item{select}{Name of the dependent variable (as string) to be used for the
test. \code{select} can also be a character vector, specifying the names of
multiple continuous variables. In this case, \code{by} is ignored and variables
specified in \code{select} are used to compute the test. This can be useful if
the data is in wide-format and no grouping variable is available.}

\item{by}{Name of the grouping variable to be used for the test. If \code{by} is
not a factor, it will be coerced to a factor. For \code{chi_squared_test()}, if
\code{probabilities} is provided, \code{by} must be \code{NULL}.}

\item{weights}{Name of an (optional) weighting variable to be used for the test.}

\item{paired}{Logical, whether to compute a paired t-test.}

\item{mu}{The hypothesized difference in means. If \code{paired = TRUE}, or for a
one-sample t-test, this is the hypothesized true mean value. If
\code{paired = FALSE}, this is the hypothesized difference in means between the
two groups.}

\item{alternative}{A character string specifying the alternative hypothesis,
must be one of \code{"two.sided"} (default), \code{"greater"} or \code{"less"}. See \code{?t.test()}.}
}
\value{
A data frame with test results.
}
\description{
This function performs a Student's t test for two independent
samples, for paired samples, or for one sample.
}
\examples{
data(efc)
# Mann-Whitney-U-Tests for elder's age by elder's sex.
t_test(efc, "e17age", by = "e16sex")

# when data is in wide-format, specify all relevant continuous
# variables in `select` and omit `by`
set.seed(123)
wide_data <- data.frame(scale1 = runif(20), scale2 = runif(20))
t_test(wide_data, select = c("scale1", "scale2"))

# same as if we had data in long format, with grouping variable
long_data <- data.frame(
  scales = c(wide_data$scale1, wide_data$scale2),
  groups = rep(c("A", "B"), each = 20)
)
t_test(long_data, select = "scales", by = "groups")
}
