% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_stan.R
\name{tidy_stan}
\alias{tidy_stan}
\title{Tidy summary output for stan models}
\usage{
tidy_stan(x, probs = 0.89, typical = "median", trans = NULL, digits = 3)
}
\arguments{
\item{x}{A \code{stanreg}, \code{stanfit} or \code{brmsfit} object.}

\item{probs}{Vector of scalars between 0 and 1, indicating the mass within
the credible interval that is to be estimated.}

\item{typical}{The typical value that will represent the Bayesian point estimate.
By default, the posterior median is returned. See \code{\link{typical_value}}
for possible values for this argument.}

\item{trans}{Name of a function or character vector naming a function, used
to apply transformations on the estimate and HDI-values. The
values for standard errors are \emph{not} transformed!}

\item{digits}{Amount of digits to round numerical values in the output.}
}
\value{
A tidy data frame, summarizing \code{x}, giving information on the
        Bayesian point estimate, standard error (which are effectively
        \emph{median absolute deviations}), HDI, ratio of effective
        sample size numbers and Rhat statistics.
}
\description{
Returns a tidy summary output for stan models.
}
\details{
Computation for HDI is based on the code from Kruschke 2015, pp. 727f.
}
\examples{
if (require("rstanarm")) {
  fit <- stan_glm(mpg ~ wt + am, data = mtcars, chains = 1)
  tidy_stan(fit)
  tidy_stan(fit, probs = c(.89, .5))
}

}
\references{
Kruschke JK. Doing Bayesian Data Analysis: A Tutorial with R, JAGS, and Stan. 2nd edition. Academic Press, 2015
}
\seealso{
\code{\link{hdi}}
}
