% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/typical.R
\name{typical_value}
\alias{typical_value}
\title{Return the typical value of a vector}
\usage{
typical_value(x, fun = c("mean", "median"))
}
\arguments{
\item{x}{A variable.}

\item{fun}{Character vector, naming the function to be applied to numeric
\code{x}. Currently, only \code{"mean"} and \code{"median"} are
supported.}
}
\value{
The "typical" value of \code{x}.
}
\description{
This function returns the "typical" value of a variable.
}
\details{
By default, for numeric variables, \code{typical_value()} returns the
         mean value of \code{x} (unless changed with the \code{fun}-argument).
         For factors, the reference level and for character vectors, to the
         most common value is returned.
}
\examples{
data(iris)
typical_value(iris$Sepal.Length)

library(purrr)
map(iris, ~ typical_value(.x))

}
