% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svy_median.R, R/wtd_mean.R, R/wtd_median.R,
%   R/wtd_sd.R, R/wtd_se.R
\name{survey_median}
\alias{survey_median}
\alias{weighted_mean}
\alias{weighted_median}
\alias{weighted_sd}
\alias{wtd_sd}
\alias{weighted_se}
\title{Weighted statistics for variables}
\usage{
survey_median(x, design)

weighted_mean(x, weights = NULL)

weighted_median(x, weights = NULL)

weighted_sd(x, weights = NULL)

wtd_sd(x, weights = NULL)

weighted_se(x, weights = NULL)
}
\arguments{
\item{x}{(Numeric) vector or a data frame. For \code{survey_median()},
the bare (unquoted) variable name, or a character vector with the variable name.}

\item{design}{An object of class \code{\link[survey]{svydesign}}, providing
a specification of the survey design.}

\item{weights}{Bare (unquoted) variable name, or a character vector with
the variable name of the numeric vector of weights. If \code{weights = NULL},
unweighted statistic is reported.}
}
\value{
The weighted (test) statistic.
}
\description{
\code{weighted_sd()}, \code{weighted_se()}, \code{weighted_mean()} and \code{weighted_median()}
compute weighted standard deviation, standard error, mean or median for a
variable or for all variables of a data frame. \code{survey_median()} computes the
median for a variable in a survey-design (see \code{\link[survey]{svydesign}}).
\code{weighted_correlation()} computes a weighted correlation for a two-sided alternative
hypothesis.
}
\note{
\code{weighted_chisq()} is a convenient wrapper for \code{\link{crosstable_statistics}}.
For a weighted one-way Anova, use \code{means_by_group()} with
\code{weights}-argument.
\cr \cr
\code{weighted_ttest()} assumes unequal variance between the two groups.
}
\examples{
# weighted sd and se ----
weighted_sd(rnorm(n = 100, mean = 3), runif(n = 100))

data(efc)
weighted_sd(efc[, 1:3], runif(n = nrow(efc)))
weighted_se(efc[, 1:3], runif(n = nrow(efc)))

# survey_median ----
# median for variables from weighted survey designs
if (require("survey")) {
  data(nhanes_sample)

  des <- svydesign(
    id = ~SDMVPSU,
    strat = ~SDMVSTRA,
    weights = ~WTINT2YR,
    nest = TRUE,
    data = nhanes_sample
  )

  survey_median(total, des)
  survey_median("total", des)
}

# weighted t-test ----
efc$weight <- abs(rnorm(nrow(efc), 1, .3))
weighted_ttest(efc, e17age, weights = weight)
weighted_ttest(efc, e17age, c160age, weights = weight)
weighted_ttest(e17age ~ e16sex + weight, efc)

# weighted Mann-Whitney-U-test ----
weighted_mannwhitney(c12hour ~ c161sex + weight, efc)

# weighted Chi-squared-test ----
weighted_chisqtest(efc, c161sex, e16sex, weights = weight, correct = FALSE)
weighted_chisqtest(c172code ~ c161sex + weight, efc)

# weighted Chi-squared-test for given probabilities ----
weighted_chisqtest(c172code ~ weight, efc, p = c(.33, .33, .34))
}
