% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{boot_ci}
\alias{boot_ci}
\alias{boot_se}
\title{Standard Error and Confidence Intervals for bootstrapped estimates}
\usage{
boot_ci(x)

boot_se(x)
}
\arguments{
\item{x}{A vector.}
}
\value{
The bootstrap standard error or the lower and upper confidence
        intervals of \code{x}.
}
\description{
Compute bootstrap standard error and confidence intervals
             for a vector of bootstrap replicate estimates.
}
\details{
This method requires a vector of bootstrap replicate estimates
         as input. The function then computes the bootstrap standard error
         by calculating the standard deviation of the input vector. The mean
         value of the input vector is used to calculate the lower and upper
         confidence interval, assuming a t-distribution of bootstrap estimate
         replicates.
}
\examples{
data(efc)
bs <- bootstrap(efc, 100)

# now run models for each bootstrapped sample
bs$models <- lapply(bs$strap, function(x) lm(neg_c_7 ~ e42dep + c161sex, data = x))

# extract coefficient "dependency" from each model
bs$dependency <- unlist(lapply(bs$models, function(x) coef(x)[2]))

# get bootstrapped confidence intervals
boot_ci(bs$dependency)

# compare with model fit
fit <- lm(neg_c_7 ~ e42dep + c161sex, data = efc)
confint(fit)[2, ]

# compare coefficients
mean(bs$dependency)
coef(fit)[2]

}

