% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chi_squared_test.R
\name{chi_squared_test}
\alias{chi_squared_test}
\title{Chi-Squared test}
\usage{
chi_squared_test(
  data,
  select = NULL,
  by = NULL,
  probabilities = NULL,
  weights = NULL,
  paired = FALSE,
  ...
)
}
\arguments{
\item{data}{A data frame.}

\item{select}{One or more name of the continuous variable (as character
vector) to be used as samples for the test. If \code{select} only specified one
variable, a one-sample test is carried out (only applicable for \code{t_test()}).
Else, \code{by} must be provided to indicate the groups of comparison.}

\item{by}{Name of the variable indicating the groups. Required if \code{select}
specifies only one variable that contains all samples to be compared in the
test. If \code{by} is not a factor, it will be coerced to a factor. For
\code{chi_squared_test()}, if \code{probabilities} is provided, \code{by} must be \code{NULL}.}

\item{probabilities}{A numeric vector of probabilities for each cell in the
contingency table. The length of the vector must match the number of cells
in the table, i.e. the number of unique levels of the variable specified
in \code{select}. If \code{probabilities} is provided, a chi-squared test for given
probabilities is conducted. Furthermore, if \code{probabilities} is given, \code{by}
must be \code{NULL}. The probabilities must sum to 1.}

\item{weights}{Name of an (optional) weighting variable to be used for the test.}

\item{paired}{Logical, if \code{TRUE}, a McNemar test is conducted for 2x2 tables.
Note that \code{paired} only works for 2x2 tables.}

\item{...}{Additional arguments passed down to \code{\link[=chisq.test]{chisq.test()}}.}
}
\value{
A data frame with test results. The returned effects sizes are
Cramer's V for tables with more than two rows and columns, Phi (\eqn{\phi})
for 2x2 tables, and \ifelse{latex}{\eqn{Fei}}{פ (Fei)} for tests against
given probabilities.
}
\description{
This function performs a \eqn{chi}^2 test for contingency
tables or tests for given probabilities. The returned effects sizes are
Cramer's V for tables with more than two rows and columns, Phi (\eqn{\phi})
for 2x2 tables, and \ifelse{latex}{\eqn{Fei}}{פ (Fei)} for tests against
given probabilities (see \emph{Ben-Shachar et al. 2023}).
}
\details{
The function is a wrapper around \code{\link[=chisq.test]{chisq.test()}} and
\code{\link[=fisher.test]{fisher.test()}} (for small expected values) for contingency tables, and
\code{chisq.test()} for given probabilities. When \code{probabilities} are provided,
these are rescaled to sum to 1 (i.e. \code{rescale.p = TRUE}). When \code{fisher.test()}
is called, simulated p-values are returned (i.e. \code{simulate.p.value = TRUE},
see \code{?fisher.test}). If \code{paired = TRUE} and a 2x2 table is provided,
a McNemar test (see \code{\link[=mcnemar.test]{mcnemar.test()}}) is conducted.

The weighted version of the chi-squared test is based on the a weighted
table, using \code{\link[=xtabs]{xtabs()}} as input for \code{chisq.test()}.
}
\examples{
\dontshow{if (requireNamespace("effectsize")) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
data(efc)
efc$weight <- abs(rnorm(nrow(efc), 1, 0.3))

# Chi-squared test
chi_squared_test(efc, "c161sex", by = "e16sex")

# weighted Chi-squared test
chi_squared_test(efc, "c161sex", by = "e16sex", weights = "weight")

# Chi-squared test for given probabilities
chi_squared_test(efc, "c161sex", probabilities = c(0.3, 0.7))
\dontshow{\}) # examplesIf}
}
\references{
Ben-Shachar, M.S., Patil, I., Thériault, R., Wiernik, B.M.,
Lüdecke, D. (2023). Phi, Fei, Fo, Fum: Effect Sizes for Categorical Data
That Use the Chi‑Squared Statistic. Mathematics, 11, 1982.
\doi{10.3390/math11091982}
}
\seealso{
\itemize{
\item \code{\link[=mann_whitney_test]{mann_whitney_test()}} for unpaired (independent) samples.
\item \code{\link[=t_test]{t_test()}} for parametric t-tests.
\item \code{\link[=kruskal_wallis_test]{kruskal_wallis_test()}} for non-parametric ANOVA (i.e. more than two samples).
\item \code{\link[=wilcoxon_test]{wilcoxon_test()}} for Wilcoxon rank sum tests for paired (dependent) samples.
\item \code{\link[=chi_squared_test]{chi_squared_test()}} for chi-squared tests (two categorical variables).
}
}
