% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chi_squared_test.R
\name{chi_squared_test}
\alias{chi_squared_test}
\title{Chi-Squared Test}
\usage{
chi_squared_test(
  data,
  select = NULL,
  by = NULL,
  probabilities = NULL,
  weights = NULL,
  ...
)
}
\arguments{
\item{data}{A data frame.}

\item{select}{Name of the dependent variable (as string) to be used for the
test.}

\item{by}{Name of the grouping variable to be used for the test. If \code{by} is
not a factor, it will be coerced to a factor. For \code{chi_squared_test()}, if
\code{probabilities} is provided, \code{by} must be \code{NULL}.}

\item{probabilities}{A numeric vector of probabilities for each cell in the
contingency table. The length of the vector must match the number of cells
in the table, i.e. the number of unique levels of the variable specified
in \code{select}. If \code{probabilities} is provided, a chi-squared test for given
probabilities is conducted. Furthermore, if \code{probabilities} is given, \code{by}
must be \code{NULL}. The probabilities must sum to 1.}

\item{weights}{Name of an (optional) weighting variable to be used for the test.}

\item{...}{Additional arguments passed down to \code{\link[=chisq.test]{chisq.test()}}.}
}
\value{
A data frame with test results.
}
\description{
This function performs a Mann-Whitney-Test (or Wilcoxon rank
sum test for \emph{unpaired} samples, see \code{\link[=wilcox.test]{wilcox.test()}} and \code{\link[coin:LocationTests]{coin::wilcox_test()}}).

The function reports p and Z-values as well as effect size r and group-rank-means.
}
\examples{
data(efc)
efc$weight <- abs(rnorm(nrow(efc), 1, 0.3))
# Chi-squared-test
chi_squared_test(efc, "c161sex", by = "e16sex")
# weighted Chi-squared-test
chi_squared_test(efc, "c161sex", by = "e16sex", weights = "weight")
# Chi-squared-test for given probabilities
chi_squared_test(efc, "c161sex", probabilities = c(0.3, 0.7))
}
