% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kruskal_wallis_test.R
\name{kruskal_wallis_test}
\alias{kruskal_wallis_test}
\title{Kruskal-Wallis-Test}
\usage{
kruskal_wallis_test(data, select = NULL, by = NULL, weights = NULL)
}
\arguments{
\item{data}{A data frame.}

\item{select}{Name of the dependent variable (as string) to be used for the
test.}

\item{by}{Name of the grouping variable to be used for the test. If \code{by} is
not a factor, it will be coerced to a factor. For \code{chi_squared_test()}, if
\code{probabilities} is provided, \code{by} must be \code{NULL}.}

\item{weights}{Name of an (optional) weighting variable to be used for the test.}
}
\value{
A data frame with test results.
}
\description{
This function performs a Kruskal-Wallis rank sum test, see
\code{\link[=kruskal.test]{kruskal.test()}} and \code{\link[coin:LocationTests]{coin::kruskal_test()}}).
}
\examples{
data(efc)
# Mann-Whitney-U-Tests for elder's age by elder's sex.
kruskal_wallis_test(efc, "e17age", by = "e16sex")
}
