% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_mann_whitney.R
\name{mann_whitney_test}
\alias{mann_whitney_test}
\title{Mann-Whitney-Test}
\usage{
mann_whitney_test(
  data,
  select = NULL,
  by = NULL,
  weights = NULL,
  distribution = "asymptotic"
)
}
\arguments{
\item{data}{A data frame.}

\item{select}{The dependent variable (numeric) to be used for the test.}

\item{by}{The grouping variable (factor) to be used for the test. If \code{by} is
not a factor, it will be coerced to a factor.}

\item{weights}{An optional weighting variable (numeric) to be used for the test.}

\item{distribution}{Indicates how the null distribution of the test statistic
should be computed. May be one of \code{"exact"}, \code{"approximate"} or \code{"asymptotic"}
(default). See \code{\link[coin:LocationTests]{coin::wilcox_test()}} for details.}
}
\value{
A data frame.
}
\description{
This function performs a Mann-Whitney-Test (or Wilcoxon rank
sum test for \emph{unpaired} samples, see \code{\link[=wilcox.test]{wilcox.test()}} and \code{\link[coin:LocationTests]{coin::wilcox_test()}}).

The function reports p and Z-values as well as effect size r and group-rank-means.
}
\details{
This function calls \code{\link[coin:LocationTests]{coin::wilcox_test()}} to extract effect sizes.
Interpretation of the effect size \strong{r}, as a rule-of-thumb:
\itemize{
\item small effect >= 0.1
\item medium effect >= 0.3
\item large effect >= 0.5
}

\strong{r} is calcuated as:

\if{html}{\out{<div class="sourceCode">}}\preformatted{r = |Z| / sqrt(n1 + n2)
}\if{html}{\out{</div>}}
}
\examples{
data(efc)
# Mann-Whitney-U-Tests for elder's age by elder's sex.
mann_whitney_test(efc, "e17age", "e16sex")
}
