% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overdisp.R
\name{overdisp}
\alias{overdisp}
\alias{zero_count}
\alias{converge_ok}
\alias{is_singular}
\alias{reliab_test}
\alias{split_half}
\alias{cronb}
\alias{difficulty}
\alias{mic}
\alias{pca}
\alias{pca_rotate}
\alias{n_eff}
\alias{mcse}
\title{Deprecated functions}
\usage{
overdisp(x, ...)

zero_count(x, ...)

converge_ok(x, ...)

is_singular(x, ...)

reliab_test(x, ...)

split_half(x, ...)

cronb(x, ...)

difficulty(x, ...)

mic(x, ...)

pca(x, ...)

pca_rotate(x, ...)

n_eff(x, ...)

mcse(x, ...)
}
\arguments{
\item{x}{An object.}

\item{...}{Currently not used.}
}
\value{
Nothing.
}
\description{
A list of deprecated functions.
}
