% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Deprecated.R
\name{overdisp}
\alias{overdisp}
\alias{zero_count}
\alias{pca}
\alias{pca_rotate}
\alias{r2}
\alias{icc}
\alias{p_value}
\alias{se}
\alias{std_beta}
\title{Deprecated functions}
\usage{
overdisp(x, ...)

zero_count(x, ...)

pca(x, ...)

pca_rotate(x, ...)

r2(x)

icc(x)

p_value(x, ...)

se(x, ...)

std_beta(x, ...)
}
\arguments{
\item{x}{An object.}

\item{...}{Currently not used.}
}
\value{
Nothing.
}
\description{
A list of deprecated functions.
}
