% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Deprecated.R
\name{overdisp}
\alias{overdisp}
\alias{zero_count}
\alias{converge_ok}
\alias{is_singular}
\alias{cronb}
\alias{pca}
\alias{pca_rotate}
\alias{n_eff}
\alias{mcse}
\alias{pred_accuracy}
\alias{r2}
\alias{icc}
\alias{hdi}
\alias{rope}
\alias{p_value}
\alias{se}
\title{Deprecated functions}
\usage{
overdisp(x, ...)

zero_count(x, ...)

converge_ok(x, ...)

is_singular(x, ...)

cronb(x, ...)

pca(x, ...)

pca_rotate(x, ...)

n_eff(x, ...)

mcse(x, ...)

pred_accuracy(x, ...)

r2(x)

icc(x)

hdi(x)

rope(x)

p_value(x, ...)

se(x, ...)
}
\arguments{
\item{x}{An object.}

\item{...}{Currently not used.}
}
\value{
Nothing.
}
\description{
A list of deprecated functions.
}
