% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Deprecated.R
\name{r2}
\alias{r2}
\alias{icc}
\alias{p_value}
\alias{se}
\alias{cohens_f}
\alias{std_beta}
\alias{robust}
\title{Deprecated functions}
\usage{
r2(x)

icc(x)

p_value(x, ...)

se(x, ...)

cohens_f(x, ...)

std_beta(x, ...)

robust(x, ...)
}
\arguments{
\item{x}{An object.}

\item{...}{Currently not used.}
}
\value{
Nothing.
}
\description{
A list of deprecated functions.
}
