% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseudo_r2.R
\name{r2}
\alias{r2}
\title{Compute R-squared of (generalized) linear (mixed) models}
\usage{
r2(x, n = NULL)
}
\arguments{
\item{x}{Fitted model of class \code{lm}, \code{glm}, \code{lmerMod}/\code{lme}
or \code{glmerMod}.}

\item{n}{Optional, a \code{lmerMod} object, representing the fitted null-model
to \code{x} (unconditional model). If \code{n} is given, the pseudo-r-squared
for random intercept and random slope variances are computed (see Kwok et al. 2008;
see 'Examples' and 'Details').}
}
\value{
\itemize{
          \item For linear models, the r-squared and adjusted r-squared values.
          \item For linear mixed models, the r-squared and Omega-squared values.
          \item For \code{glm} objects, Cox & Snell's and Nagelkerke's pseudo r-squared values.
          \item For \code{glmerMod} objects, Tjur's coefficient of determination.
        }
}
\description{
Compute R-squared values of linear (mixed) models, or
               pseudo-R-squared values for generalized linear (mixed) models.
}
\details{
If \code{n} is given, the Pseudo-R2 statistic is the proportion of
         explained variance in the random effect after adding co-variates or
         predictors to the model, or in short: the proportion of the explained
         variance in the random effect of the full (conditional) model \code{x}
         compared to the null (unconditional) model \code{n}.
}
\note{
For linear models, the r-squared and adjusted r-squared value is returned,
        as provided by the \code{summary}-function.
        \cr \cr
        For linear mixed models, an r-squared approximation by computing the
        correlation between the fitted and observed values, as suggested by
        Byrnes (2008), is returned as well as the Omega-squared value as
        suggested by Xu (2003), unless \code{n} is specified. If \code{n}
        is given, pseudo r-squared measures based on the variances of random
        intercept (tau 00, between-group-variance) and random slope (tau 11,
        random-slope-variance) are returned.
        \cr \cr
        For generalized linear models, Cox & Snell's and Nagelkerke's
        pseudo r-squared values are returned.
        \cr \cr
        For generalized linear mixed models, the coefficient of determination
        as suggested by Tjur (2009) (see also \code{\link{cod}}).
}
\examples{
library(sjmisc)
library(lme4)
fit <- lmer(Reaction ~ Days + (Days | Subject), sleepstudy)
r2(fit)

data(efc)
fit <- lm(barthtot ~ c160age + c12hour, data = efc)
r2(fit)

# Pseudo-R-squared values
efc$services <- ifelse(efc$tot_sc_e > 0, 1, 0)
fit <- glm(services ~ neg_c_7 + c161sex + e42dep,
           data = efc, family = binomial(link = "logit"))
r2(fit)

# Pseudo-R-squared values for random effect variances
fit <- lmer(Reaction ~ Days + (Days | Subject), sleepstudy)
fit.null <- lmer(Reaction ~ 1 + (Days | Subject), sleepstudy)
r2(fit, fit.null)


}
\references{
\itemize{
              \item \href{http://glmm.wikidot.com/faq}{DRAFT r-sig-mixed-models FAQ}
              \item Byrnes, J. 2008. Re: Coefficient of determination (R^2) when using lme() (\url{https://stat.ethz.ch/pipermail/r-sig-mixed-models/2008q2/000713.html})
              \item Kwok OM, Underhill AT, Berry JW, Luo W, Elliott TR, Yoon M. 2008. Analyzing Longitudinal Data with Multilevel Models: An Example with Individuals Living with Lower Extremity Intra-Articular Fractures. Rehabilitation Psychology 53(3): 370–86. \doi{10.1037/a0012765}
              \item Xu, R. 2003. Measuring explained variation in linear mixed effects models. Statist. Med. 22:3527-3541. \doi{10.1002/sim.1572}
              \item Tjur T. 2009. Coefficients of determination in logistic regression models - a new proposal: The coefficient of discrimination. The American Statistician, 63(4): 366-372
            }
}

