% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icc.R
\name{re_var}
\alias{re_var}
\alias{get_re_var}
\title{Random effect variances}
\usage{
re_var(x)

get_re_var(x, comp = c("tau.00", "tau.01", "tau.11", "rho.01", "sigma_2"))
}
\arguments{
\item{x}{Fitted mixed effects model (of class \code{\link[lme4]{merMod}} or
\code{\link[glmmTMB]{glmmTMB}}). \code{get_re_var()} also accepts
an object of class \code{icc.lme4}, as returned by the
\code{\link{icc}} function.}

\item{comp}{Name of the variance component to be returned. See 'Details'.}
}
\value{
\code{get_re_var()} returns the value of the requested variance component,
          \code{re_var()} returns \code{NULL}.
}
\description{
These functions extracts random effect variances as well as
               random-intercept-slope-correlation of mixed effects models.
               Currently, \code{\link[lme4]{merMod}} and
               \code{\link[glmmTMB]{glmmTMB}} objects are supported.
}
\details{
The random effect variances indicate the between- and within-group
        variances as well as random-slope variance and random-slope-intercept
        correlation. Use following values for \code{comp} to get the particular
        variance component:
        \describe{
         \item{\code{"sigma_2"}}{Within-group (residual) variance}
         \item{\code{"tau.00"}}{Between-group-variance (variation between individual intercepts and average intercept)}
         \item{\code{"tau.11"}}{Random-slope-variance (variation between individual slopes and average slope)}
         \item{\code{"tau.01"}}{Random-Intercept-Slope-covariance}
         \item{\code{"rho.01"}}{Random-Intercept-Slope-correlation}
        }
}
\examples{
library(lme4)
fit1 <- lmer(Reaction ~ Days + (Days | Subject), sleepstudy)

# all random effect variance components
re_var(fit1)

# just the rand. slope-intercept covariance
get_re_var(fit1, "tau.01")

sleepstudy$mygrp <- sample(1:45, size = 180, replace = TRUE)
fit2 <- lmer(Reaction ~ Days + (1 | mygrp) + (Days | Subject), sleepstudy)
re_var(fit2)

}
\references{
Aguinis H, Gottfredson RK, Culpepper SA. 2013. Best-Practice Recommendations for Estimating Cross-Level Interaction Effects Using Multilevel Modeling. Journal of Management 39(6): 1490–1528 (\doi{10.1177/0149206313478188})
}
\seealso{
\code{\link{icc}}
}
