% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/re-exports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{mse}
\alias{rmse}
\alias{link_inverse}
\alias{weighted_sd}
\alias{weighted_mean}
\alias{weighted_median}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{datawizard}{\code{\link[datawizard]{weighted_mean}}, \code{\link[datawizard:weighted_mean]{weighted_median}}, \code{\link[datawizard:weighted_mean]{weighted_sd}}}

  \item{insight}{\code{\link[insight]{link_inverse}}}

  \item{performance}{\code{\link[performance:performance_mse]{mse}}, \code{\link[performance:performance_rmse]{rmse}}}
}}

