% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/re-exports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{\%>\%}
\alias{typical_value}
\alias{r2}
\alias{icc}
\alias{mse}
\alias{performance_rse}
\alias{rmse}
\alias{hdi}
\alias{ci}
\alias{rope}
\alias{equivalence_test}
\alias{link_inverse}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{bayestestR}{\code{\link[bayestestR]{hdi}}, \code{\link[bayestestR]{ci}}, \code{\link[bayestestR]{rope}}, \code{\link[bayestestR]{equivalence_test}}}

  \item{insight}{\code{\link[insight]{link_inverse}}}

  \item{magrittr}{\code{\link[magrittr]{\%>\%}}}

  \item{performance}{\code{\link[performance]{r2}}, \code{\link[performance]{icc}}, \code{\link[performance]{mse}}, \code{\link[performance]{performance_rse}}, \code{\link[performance]{rmse}}}

  \item{sjmisc}{\code{\link[sjmisc]{typical_value}}}
}}

