\encoding{UTF-8}
\name{sjstats-package}
\alias{sjstats-package}
\alias{sjstats}
\docType{package}
\title{Statistical Functions for Regression Models}
\description{
'sjstats' aims at providing convenient functions for common statistical computations, which are not provided by R's base or stats packages. This package comprises tools that help computing various statistics for simple regression or mixed models. These include:
\itemize{
  \item For regression and mixed models: Coefficient of Variation, Root Mean Squared Error, Residual Standard Error, Coefficient of Discrimination, R-squared and pseudo-R-squared values, standardized beta values
  \item Especially for mixed models: Design effect, ICC, sample size calculation, convergence and overdispersion tests
}
Some functions are also for simple statistics, like:
\itemize{
  \item Cramer's V, Cronbach's Alpha, Mean Inter-Item-Correlation, Mann-Whitney-U-Test, Item-scale reliability tests
}
}
\details{
\tabular{rl}{
Package: \tab sjstats\cr
Type: \tab Package\cr
Version: \tab 0.1.0\cr
Date: \tab 2016-05-20\cr
License: \tab GPL-3\cr
}
}
\author{
Daniel Lüdecke \email{d.luedecke@uke.de}
}
\references{
  \itemize{
    \item Documentation: \url{http://www.strengejacke.de/sjPlot}
    \item Github: \url{https://github.com/sjPlot/sjstats}
    \item Bug reports: \url{https://github.com/sjPlot/sjstats/issues}
    }
}
