% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_test.R
\name{t_test}
\alias{t_test}
\title{Student's t test}
\usage{
t_test(
  data,
  select = NULL,
  by = NULL,
  weights = NULL,
  paired = FALSE,
  mu = 0,
  alternative = "two.sided"
)
}
\arguments{
\item{data}{A data frame.}

\item{select}{One or more name of the continuous variable (as character
vector) to be used as samples for the test. If \code{select} only specified one
variable, a one-sample test is carried out (only applicable for \code{t_test()}).
Else, \code{by} must be provided to indicate the groups of comparison.}

\item{by}{Name of the variable indicating the groups. Required if \code{select}
specifies only one variable that contains all samples to be compared in the
test. If \code{by} is not a factor, it will be coerced to a factor. For
\code{chi_squared_test()}, if \code{probabilities} is provided, \code{by} must be \code{NULL}.}

\item{weights}{Name of an (optional) weighting variable to be used for the test.}

\item{paired}{Logical, whether to compute a paired t-test.}

\item{mu}{The hypothesized difference in means (for \code{t_test()}) or location
shift (for \code{mann_whitney_test()}). The default is 0.}

\item{alternative}{A character string specifying the alternative hypothesis,
must be one of \code{"two.sided"} (default), \code{"greater"} or \code{"less"}. See \code{?t.test}
and \code{?wilcox.test}.}
}
\value{
A data frame with test results.
}
\description{
This function performs a Student's t test for two independent
samples, for paired samples, or for one sample. Unlike the underlying
base R function \code{t.test()}, this function allows for weighted tests and
automatically calculates effect sizes.
}
\examples{
\dontshow{if (requireNamespace("effectsize")) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
data(sleep)
# one-sample t-test
t_test(sleep, "extra")
# base R equivalent
t.test(extra ~ 1, data = sleep)

# two-sample t-test, by group
t_test(mtcars, "mpg", by = "am")
# base R equivalent
t.test(mpg ~ am, data = mtcars)

# paired t-test
t_test(mtcars, c("mpg", "hp"), paired = TRUE)
# base R equivalent
t.test(mtcars$mpg, mtcars$hp, data = mtcars, paired = TRUE)
\dontshow{\}) # examplesIf}
}
