% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_stan.R
\name{tidy_stan}
\alias{tidy_stan}
\title{Tidy summary output for stan models}
\usage{
tidy_stan(x, probs = 0.89, typical = "median", trans = NULL,
  type = c("fixed", "random", "all"), digits = 3)
}
\arguments{
\item{x}{A \code{stanreg}, \code{stanfit} or \code{brmsfit} object.}

\item{probs}{Vector of scalars between 0 and 1, indicating the mass within
the credible interval that is to be estimated. See \code{\link{hdi}}.}

\item{typical}{The typical value that will represent the Bayesian point estimate.
By default, the posterior median is returned. See \code{\link{typical_value}}
for possible values for this argument.}

\item{trans}{Name of a function or character vector naming a function, used
to apply transformations on the estimate and HDI-values. The
values for standard errors are \emph{not} transformed!}

\item{type}{For mixed effects models, specify the type of effects that should
be returned. \code{type = "fixed"} returns fixed effects only,
\code{type = "random"} the random effects and \code{type = "all"} returns
both fixed and random effects.}

\item{digits}{Amount of digits to round numerical values in the output.}
}
\value{
A tidy data frame, summarizing \code{x}, with consistent column names.
        To distinguish multiple HDI values, column names for the HDI get a suffix
        when \code{probs} has more than one element.
}
\description{
Returns a tidy summary output for stan models.
}
\details{
The returned data frame gives information on the Bayesian point
         estimate (column \emph{estimate}, which is by default the posterior
         median; other statistics are also possible, see \code{typical}), the
         standard error (which are actually \emph{median absolute deviations}),
         the HDI, the ratio of effective numbers of samples, \emph{n_eff},
         (i.e. effective number of samples divided by total number of samples)
         and Rhat statistics.
         \cr \cr
         The ratio of effective number of samples ranges from 0 to 1,
         and should be close to 1. The closer this ratio comes to zero means
         that the chains may be inefficient, but possibly still okay.
         \cr \cr
         When Rhat is above 1, it usually indicates that the chain has not
         yet converged, indicating that the drawn samples might not be
         trustworthy. Drawing more iteration may solve this issue.
         \cr \cr
         Computation for HDI is based on the code from Kruschke 2015, pp. 727f.
}
\examples{
\dontrun{
if (require("rstanarm")) {
  fit <- stan_glm(mpg ~ wt + am, data = mtcars, chains = 1)
  tidy_stan(fit)
  tidy_stan(fit, probs = c(.89, .5))
}}

}
\references{
Kruschke JK. \emph{Doing Bayesian Data Analysis: A Tutorial with R, JAGS, and Stan.} 2nd edition. Academic Press, 2015
\cr \cr
Gelman A, Carlin JB, Stern HS, Dunson DB, Vehtari A, Rubin DB. \emph{Bayesian data analysis.} 3rd ed. Boca Raton: Chapman & Hall/CRC, 2013
\cr \cr
Gelman A, Rubin DB. \emph{Inference from iterative simulation using multiple sequences.} Statistical Science 1992;7: 457–511
\cr \cr
McElreath R. \emph{Statistical Rethinking. A Bayesian Course with Examples in R and Stan.} Chapman and Hall, 2015
}
\seealso{
\code{\link{hdi}}
}
